refactor: put reports in directory, move code to file
